evaluation/requirements.txt
google-api-python-client==2.151.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
streamlit==1.37.0
gspread==6.1.2
deepeval==1.4.9
langchain-google-vertexai==2.0.6
asyncio==3.4.3
huggingface-hub==0.26.2
